correccion en inventario fisico ingresos y egresos cantidad 0

+++ application/wms/models/Ingreso_model.php
@@ -95,7 +95,7 @@ class Ingreso_model extends General_Model {
 				
 				if (verDato($args, "_costo")) {
 					$row->precio_total = round($row->precio_total + $row->precio_costo_iva, 2);
-					$row->precio_unitario = round((int)$row->cantidad !== 0 ? ($row->precio_total / $row->cantidad) : 0.00, 2);
+					$row->precio_unitario = round((float)$row->cantidad !== 0 ? ($row->precio_total / $row->cantidad) : 0.00, 2);
 				}
 				$datos[] = $row;
 			}
